backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
alembic==1.13.1

# LLM and AI
openai==1.12.0
langchain==0.1.5
langgraph==0.0.20
langchain-openai==0.0.5
# chromadb==0.4.22  # Optional - requires C++ build tools on Windows. Use Docker for production.

# Document Processing
pypdf2==3.0.1
reportlab==4.0.9
pillow==10.2.0
python-docx==1.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2024.1

# Monitoring and Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

